+++ changelog.md
@@ -1,6 +1,7 @@
 # FluxLoader v0.10.0
 - Added `onLuaFilesLoaded` and `onLuaScriptExecuted` events
 - Fixed names of methods for exposing classes and objects
+- Fixed log output when creating backups and restoring
 
 # FluxLoader v0.9.0
 - Added methods for getting and setting lists of specific types for plugin configs